requirements.txt
python-keystoneclient>=0.10.0
cliff>=1.4.3
